editing

+++ dating_site/dating_site/profileupdate_info.aspx.cs
@@ -18,52 +18,67 @@ namespace dating_site
         protected void register_register_Click(object sender, EventArgs e)
         {
 
-            try
+            if (!FileUpload1.HasFile)
             {
-                int my_id = (int)Session["id"];
-
-                string filtered_email = Session["email"].ToString();
-                int index = filtered_email.IndexOf('@');
-                filtered_email = filtered_email.Substring(0, index);
-
-
-              
-                string strname = FileUpload1.FileName.ToString();
-                var result = strname.Substring(strname.Length - 4);
-
-                string filtered_name = filtered_email + result; ;
-
-                File.Delete(Server.MapPath("~/upload/") + filtered_name);
-
-                FileUpload1.PostedFile.SaveAs(Server.MapPath("~/upload/") + filtered_name);
-
-                filtered_name = "~/upload/" + filtered_name.Trim(); ;
-
-
-                System.Data.SqlClient.SqlConnection sqlConnection1 = new System.Data.SqlClient.SqlConnection("Data Source = uvuserdata.mssql.somee.com; Initial Catalog = uvuserdata; Persist Security Info = True; User ID = yuvrajvirk55_SQLLogin_1; Password = nm6ecevlt8");
-                System.Data.SqlClient.SqlCommand cmd = new System.Data.SqlClient.SqlCommand();
-
-                cmd.CommandType = System.Data.CommandType.Text;
-                cmd.CommandText = "UPDATE usertable SET first_name = '"+ register_fname.Text + "', last_name = '" + register_lname.Text + "', dob = '" + register_dob.Text + "', sex = '" + register_gender.Text + "', pic = '" +filtered_name + "' where id = " +my_id + "";
-                cmd.Connection = sqlConnection1;
-                
-                sqlConnection1.Open();
-                cmd.ExecuteNonQuery();
-
-                sqlConnection1.Close();
-
-         
-                Session["Name"] = register_fname.Text+ " "+ register_lname.Text;
-                Session["img"] = filtered_name.ToString();
-                Response.Redirect("profile.aspx");
-
-
+                Page.ClientScript.RegisterClientScriptBlock(typeof(Page), "Alert", "alert('Upload image')", true);
             }
 
-            catch
+            else if (FileUpload1.FileName.ToString().Contains(".jpeg"))
+            {
+                Page.ClientScript.RegisterClientScriptBlock(typeof(Page), "Alert", "alert('Upload Different Format')", true);
+            }
+            else
             {
 
-                Page.ClientScript.RegisterClientScriptBlock(typeof(Page), "Alert", "alert('Info Change Failed!')", true);
+                try
+                {
+                    int my_id = (int)Session["id"];
+
+                    string filtered_email = Session["email"].ToString();
+                    int index = filtered_email.IndexOf('@');
+                    filtered_email = filtered_email.Substring(0, index);
+
+
+
+                    string strname = FileUpload1.FileName.ToString();
+                    var result = strname.Substring(strname.Length - 4);
+
+                    string filtered_name = filtered_email + result; ;
+
+                    File.Delete(Server.MapPath("~/upload/") + filtered_name);
+
+                    FileUpload1.PostedFile.SaveAs(Server.MapPath("~/upload/") + filtered_name);
+
+                    filtered_name = "~/upload/" + filtered_name.Trim(); ;
+
+
+
+
+                    System.Data.SqlClient.SqlConnection sqlConnection1 = new System.Data.SqlClient.SqlConnection("Data Source = uvuserdata.mssql.somee.com; Initial Catalog = uvuserdata; Persist Security Info = True; User ID = yuvrajvirk55_SQLLogin_1; Password = nm6ecevlt8");
+                    System.Data.SqlClient.SqlCommand cmd = new System.Data.SqlClient.SqlCommand();
+
+                    cmd.CommandType = System.Data.CommandType.Text;
+                    cmd.CommandText = "UPDATE usertable SET first_name = '" + register_fname.Text + "', last_name = '" + register_lname.Text + "', dob = '" + register_dob.Text + "', sex = '" + register_gender.Text + "', pic = '" + filtered_name + "' where id = " + my_id + "";
+                    cmd.Connection = sqlConnection1;
+
+                    sqlConnection1.Open();
+                    cmd.ExecuteNonQuery();
+
+                    sqlConnection1.Close();
+
+
+                    Session["Name"] = register_fname.Text + " " + register_lname.Text;
+                    Session["img"] = filtered_name.ToString();
+                    Response.Redirect("profile.aspx");
+
+
+                }
+
+                catch
+                {
+
+                    Page.ClientScript.RegisterClientScriptBlock(typeof(Page), "Alert", "alert('Info Change Failed!')", true);
+                }
             }
         }
     }
